refactor: split CodeMirrorDiffView into focused modules
- review-extensions.ts: pure CM state field, effect, gutter markers
- MergeViewEditor.tsx: CM lifecycle, comment gutter, hover, portals
- CodeMirrorDiffView.tsx: container with file loading and scroll tracking
